@@ -1,4 +1,4 @@
-name: flutter_appskeleton
+name: flutter_skeletonapp
 description: A new Flutter application.
 
 # The following line prevents the package from being accidentally published to
@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -27,18 +27,27 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  toast: ^0.1.5
-  shared_preferences: ^0.5.10
-  flutter_svg: ^0.18.1
-  flutter_screenutil: ^2.3.1
-  dio: ^3.0.10
-  connectivity: ^0.4.9+2
-  cached_network_image: ^2.3.2+1
-  flutter_device_type: ^0.2.0
+  cupertino_icons: ^1.0.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  cupertino_icons: ^1.0.3
+  shared_preferences: ^2.0.5
+  flutter_svg: ^0.22.0
+  flutter_screenutil: ^5.0.0+2
+  dio: ^4.0.0
+  connectivity: ^3.0.3
+  flutter_swiper: ^1.1.6
+  flutter_device_type: ^0.4.0
+  cached_network_image: ^3.0.0
+  firebase_messaging: ^10.0.0
+  webview_flutter: ^1.0.7
+  flutter_local_notifications: ^5.0.0
+  rxdart: ^0.27.0
+  shimmer: ^2.0.0
+  intl: ^0.17.0
+
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
@@ -51,25 +60,29 @@ flutter:
   # the material Icons class.
   uses-material-design: true
   assets:
-  - assets/images/
-  - assets/
+    - assets/images/
+    - assets/
 
   fonts:
-  - family: Reguler
-    fonts:
-    - asset: assets/fonts/font-Regular.ttf
-  - family: Medium
-    fonts:
-    - asset: assets/fonts/font-Medium.ttf
-  - family: Bold
-    fonts:
-    - asset: assets/fonts/font-Bold.ttf
-  - family: Italic
-    fonts:
-    - asset: assets/fonts/font-Italic.ttf
-  - family: Light
-    fonts:
-    - asset: assets/fonts/font-Light.ttf
+    - family: Reguler
+      fonts:
+        - asset: assets/fonts/font-Regular.ttf
+    - family: Medium
+      fonts:
+        - asset: assets/fonts/font-Medium.ttf
+    - family: Bold
+      fonts:
+        - asset: assets/fonts/font-Bold.ttf
+    - family: Italic
+      fonts:
+        - asset: assets/fonts/font-Italic.ttf
+    - family: Light
+      fonts:
+        - asset: assets/fonts/font-Light.ttf
+    - family: SemiBold
+      fonts:
+        - asset: assets/fonts/font-SemiBold.ttf
+
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
