@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 
 /// Wrapper for stateful functionality to provide onInit calls in stateles widget
 class StatefulWrapper extends StatefulWidget {
-  final Function onInit;
-  final Widget child;
+  final Function? onInit;
+  final Widget? child;
   const StatefulWrapper({@required this.onInit, @required this.child});
   @override
   _StatefulWrapperState createState() => _StatefulWrapperState();
@@ -13,13 +13,13 @@ class _StatefulWrapperState extends State<StatefulWrapper> {
   @override
   void initState() {
     if (widget.onInit != null) {
-      widget.onInit();
+      widget.onInit!();
     }
     super.initState();
   }
 
   @override
   Widget build(BuildContext context) {
-    return widget.child;
+    return widget.child!;
   }
 }
