@@ -14,12 +14,15 @@ import 'package:flutter/material.dart';
 
 class Resume {
   dynamic data;
-  String source;
+  String? source;
 }
 
 abstract class ResumableState<T extends StatefulWidget> extends State with WidgetsBindingObserver {
   Resume resume = new Resume();
   bool _isPaused = false;
+BuildContext maintcontext;
+
+  ResumableState(this.maintcontext);
 
   /// Implement your code here
   void onResume() {
@@ -38,7 +41,7 @@ abstract class ResumableState<T extends StatefulWidget> extends State with Widge
   }
 
   /// This method is replacement of Navigator.push(), but fires onResume() after route popped
-  Future<T> push<T extends Object>(BuildContext context, Route<T> route, [String source]) {
+  Future<T> push<T extends Object>(BuildContext context, Route<T> route, [String? source]) {
     _isPaused = true;
     onPause();
 
@@ -49,12 +52,12 @@ abstract class ResumableState<T extends StatefulWidget> extends State with Widge
       resume.source = source;
 
       onResume();
-      return value;
+      return value!;
     });
   }
 
   /// This method is replacement of Navigator.pushNamed(), but fires onResume() after route popped
-  Future<T> pushNamed<T extends Object>(BuildContext context, String routeName, {Object arguments}) {
+  Future<T> pushNamed<T extends Object>(BuildContext context, String routeName, {Object? arguments}) {
     _isPaused = true;
     onPause();
 
@@ -65,7 +68,7 @@ abstract class ResumableState<T extends StatefulWidget> extends State with Widge
       resume.source = routeName;
 
       onResume();
-      return value;
+      return value!;
     });
   }
 
@@ -73,15 +76,15 @@ abstract class ResumableState<T extends StatefulWidget> extends State with Widge
   void initState() {
     super.initState();
 
-    WidgetsBinding.instance.addObserver(this);
-    WidgetsBinding.instance.addPostFrameCallback((_) => onReady());
+    WidgetsBinding.instance!.addObserver(this);
+    WidgetsBinding.instance!.addPostFrameCallback((_) => onReady());
 
   }
 
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
     super.dispose();
   }
 
