@@ -1,6 +1,21 @@
 class StringConstants{
  static const String PREF_TOKEN="PREF_TOKEN";
  static const String PREF_USERID="PREF_USERID";
- static const String PREF_ACHIVEID="PREF_ACHIVEID";
+ static const String PREF_FAMILYID="PREF_FAMILYID";
+ static const String PREF_USERNAME="PREF_USERNAME";
+ static const String PREF_EMAIL="PREF_EMAIL";
+ static const String PREF_SCREENSETTINGS="PREF_SCREENSETTINGS";
+ static const String PREF_PLAYBACKSETTINGS="PREF_PLAYBACKSETTINGS";
+
+  static const String COMMON_DATEFORMAT="yyyy-MM-dd HH:mm:ss";
+
+  static const String NOTIFICATION_CHANELID="SomethingNew_ChatRequest";
+  static const String NOTIFICATION_CHANELNAME="SomethingNew_ChatRequest";
+  static const String NOTIFICATION_CHANELDESCRIPTION="SomethingNew app giving to ChatRequest";
+
+  static const String MAINFOLDER="seesay";
+  static const String IMAGEFOLDER="images";
+  static const String AUDIOFOLDER="audios";
+
 
 }
