@@ -1,18 +1,35 @@
-import 'package:flutter_appskeleton/customtexts/AppStrings.dart';
+class UrlConstants {
 
-class UrlConstants{
 
- static const String Baseurl="https://uptodatefi.com/api";
+//local server
+ //static const String Baseurl = "http://192.168.1.31:8000/api";
+ // static const String imagurl = "http://192.168.1.31:8000";
+  // static const String staticpageurl = "http://192.168.1.31:8000/api";
 
-static const  String LOGIN="/auth/applogin";
-static const  String SIGNUP="/auth/signup";
-static const  String USERINTERLIST="/planachieve";
-static const  String UPDATEUSERINTREST="/userInterest";
 
-//Request Code
-static const  int REG_LOGIN=101;
-static const  int REG_SIGNUP=102;
-static const  int REG_USERINTERLIST=103;
-static const  int REG_UPDATEUSERINTREST=104;
+//live server
+static const String Baseurl = "http://appdevelopmentsingapore.com/flutter_skeletonapp/api";
+static const String imagurl = "http://appdevelopmentsingapore.com/seesay";
+static const String staticpageurl = "http://appdevelopmentsingapore.com/flutter_skeletonapp/api";
+
+static const String signup = "/signup";
+static const String login = "/login";
+static const String logout = "/logout";
+static const String savealbum = "/saveAlbum";
+static const String editalbum = "/editAlbum";
+static const String listalbum = "/listAlbum";
+static const String settings = "/settings/";
+static const String savefavouritealbum = "/saveFavourite";
+static const String removefavouritealbum = "/removeFavourite";
+static const String forgetpassword = "/forgotPassword";
+static const String forgotPasswordVerifyOtp = "/forgotPasswordVerifyOtp";
+static const String forgotResetPassword = "/forgotResetPassword";
+static const String savecontact = "/saveContact";
+static const String deletephoto = "/deletePhoto";
+
+
+//static pages
+  static const String howtouseseesay ="${staticpageurl}/howToUse";
+  static const String aboutseesay ="${staticpageurl}/about";
 
 }
