@@ -1,54 +1,71 @@
-class AppStrings{
-
-  static const String languagejson='{"appname":"UpToFi",'
-      '"privacytext":"By continuing you agree that you have read and accept our",'
-      '"copyrights":"© 2020 Dr. Khubybe’s Ltd. All rights reserved",'
-      '"copyrightslocation":"Made with love in London",'
-      '"tellasabout":"Tell us about your interest?",'
-      '"emailhint":"E-mail",'
-      '"passwordhint":"Password",'
-      '"Appname":"UpToDateFi",'
-      '"appnamehint":"scientific Research, Truly Up To Date and Simplified",'
-      '"loginwithemail":"Login With Email",'
-      '"login":"LOGIN",'
-      '"loginnow":"LOGIN NOW",'
-      '"forgotpassword":"FORGOT PASSWORD ?",'
-      '"rights":"© 2020 Dr. Khubybe’s Ltd. All rights reserved",'
-      '"rightslocation":"Made with love in London",'
-      '"unkownerror":"Some thing went wrong",'
-      '"tellusintrest":"Tell us about your interest?",'
-      '"Ready":"Ready!",'
-      '"startjourney":"Start Your Journey",'
-      '"letsstart":"LET\'S START",'
+class AppStrings {
+  static const String languagejson = '{"appname":"Seesay",'
+      '"Back":"Back",'
+      '"getstarted":"Get Started",'
+      '"editalbum":"      Edit Album",'
+      '"editaalbum":"Edit Album",'
+      '"careatalbum":"Create Your Own Album",'
+      '"tochangepassword":"Forget your password?",'
+      '"verifyoptp":"Please verify OTP",'
+      '"Newpassword":"New Password",'
+      '"confrimpassword":"Confirm Password",'
+      '"changenewpass":"Enter New Password",'
+      '"changepassword":"Change Password",'
+      '"otp":"OTP",'
+      '"confrimotp":"Confirm OTP",'
+      '"getotp":"Submit",'
+      '"Remove photo":" Remove photo",'
+      '"albumtitle":"Album Title",'
+      '"maximumcharacters":"Maximum = 22 characters",'
+      '"addphoto":"Add Photo(s)",'
+      '"flashcard":"Flashcard Word",'
+      '"albumcover":"Album Cover ?",'
+      '"albumcoverno":"Album Cover",'
+      '"imagetext":"Image Text",'
+      '"recording&playback":"Recording and Playback",'
+      '"examplefood":"Example: Food ",'
+      '"subscription":"Subscription ",'
+      '"freeplan":"Free Plan ",'
+      '"upgrade":"Upgrade",'
+      '"accountsharing":"Account Sharing",'
+      '"8Xremaining":"You have 8x uploads remaining",'
+      '"regpasstext":"To change the  password, Please provide your registered Email id.",'
+      '"regOtptext":"Please check your Email id for Otp.",'
+      '"yoursubcription":"Your Subscription",'
+      '"editmyfavourite":"Edit My Favourites Album",'
+      '"settings":"Settings",'
+      '"on":"On",'
+      '"off":"Off",'
+      '"landscapemode":"Force Landscape Mode",'
+      '"onesecselay":"1 second delay before playback (default)",'
+      '"threesecselay":"3 second delay before playback.",'
+      '"delaytapped":"Playback only when tapped.",'
+      '"star":"Star ",'
+      '"validonly":"OTP is valid only for 15 minutes.",'
+      '"playbacksettings":"Playback Settings",'
+      '"addtomyfavouritealbum":" flashcards when viewing to add them to My Favourites Album.",'
       '"next":"Next",'
-      '"Home":"Home",'
-      '"Library":"Library",'
-      '"Explore":"Explore",'
-      '"Repetition":"Repetition",'
-      '"Me":"Me",'
-      '"whatsnew":"What is new!",'
-      '"pickedforyou":"Picked for you!",'
-      '"careersuccess":"Career & Success",'
-      '"communication":"Communication(Skills)",'
-      '"creativity":"Creativity",'
-      '"economics":"Economics",'
-      '"education":"Education",'
-      '"entrepreneurship":"Entrepreneurship",'
-      '"happiness":"Happiness",'
-      '"innovation":"Innovation",'
-      '"leadership":"Leadership",'
-      '"management":"Management",'
-      '"marketing":"Marketing",'
-      '"medical":"Medical",'
-      '"mindbrain":"Mind & Brain",'
-      '"nature":"Nature & Environment",'
-      '"personlGrowth":"Personal Growth & Motivation",'
-      '"Politics":"Politics",'
-      '"relationship":"Relationship",'
-      '"society":"Society & Culture",'
-      '"technology":"Technology"'
-  '}';
-
-
-
+      '"login":"Login",'
+      '"email":"Email Address",'
+      '"login":"Login",'
+      '"confrim":"Confirm",'
+      '"password":"Password",'
+      '"spouseemail":"Spouse Email Address",'
+      '"fullname":"Full Name",'
+      '"takephoto":"Take a Photo",'
+      '"logout":"Logout",'
+      '"contactus":"Contact Us",'
+      '"howtouse":"How to Use See-Say ",'
+      '"submit":"Submit ",'
+      '"title":"Title",'
+      '"description":"Description",'
+      '"newaccount":"Don’t have an account ? ",'
+      '"signup":"Signup",'
+      '"forgetpassword":"Forget Password?",'
+      '"addfromdevice":"Add from Device",'
+      '"connectionerror":"connect your internet connection",'
+      '"unkownerror":"Something went wrong",'
+      '"connectiontimeerror":"Connection time out",'
+      '"secondemail":"Second Email Address (optional)"'
+      '}';
 }
