@@ -1,18 +1,21 @@
 
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-class dimension{
+class Dimension{
 
 
 
-  static const double text_extra_very_small=8;
-  static const double text_very_small=10;
-  static const double text_small=12;
-  static const double text_medium=14;
-  static const double text_large=16;
-  static const double text_very_large=18;
-  static const double text_mega=20;
-  static const double text_twintyfive=25;
-  static const double very_very_mega=40;
+
+  static  double get text_extra_very_small=>ScreenUtil().setSp(8);
+  static  double get text_very_small=>ScreenUtil().setSp(10);
+  static double get textSmall =>ScreenUtil().setSp(12);
+  static  double get text_medium=>ScreenUtil().setSp(14);
+  static  double get text_large=>ScreenUtil().setSp(16);
+  static  double get text_very_large=>ScreenUtil().setSp(18);
+  static  double get text_mega=>ScreenUtil().setSp(20);
+  static  double get text_twintyfive=>ScreenUtil().setSp(25);
+  static  double get very_very_mega=>ScreenUtil().setSp(40);
+
 
 
 
